Add IPK packing, CRC, compression, and test suite
- Implement JustDanceIPKWriter for packing directories into IPK files with compression and CRC calculation
- Add UbiArtCRC and Compressor utilities for CRC and zlib-compliant compression
- Extend BinaryReader/BinaryWriter with big-endian and NT string support
- Refactor Program.cs for unified pack/extract CLI with improved help and error handling
- Add comprehensive xUnit test project covering CRC, compression, endianness, and full pack/unpack integration

+++ TextureConverter/TextureType/XTX.cs
@@ -412,7 +412,7 @@ public class XTX
 
         writer.BaseStream.Seek(headerStartPos + 4 + 4, SeekOrigin.Begin);
         writer.Write((ulong)estimatedDataSize);
-        writer.Write((long)actualDataOffset);
+        writer.Write(actualDataOffset);
 
         writer.BaseStream.Seek(endPos, SeekOrigin.Begin);
     }
